# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a high-performance dating app backend server.",
        "title": "Dating App API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/swipe": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Records a user's swipe (LIKE, DISLIKE, or CRUSH) on a target entity and checks for a mutual match.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "swipe"
                ],
                "summary": "Record a swipe",
                "parameters": [
                    {
                        "description": "Swipe action details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swipe.SwipeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Swipe recorded result",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/swipe.MatchResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/swipe/candidates": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetches a list of potential entities (solo users or groups) based on entity type and user preferences.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "swipe"
                ],
                "summary": "Get swipe candidates",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "distance",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "\"user\" or \"group\", empty = all",
                        "name": "entity_type",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "genders",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "interests",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "latitude",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "longitude",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "max_age",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "max_height",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "min_age",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "min_height",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "relationship_types",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "swiper_entity_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Candidate list",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/response.EntityResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/swipe/likes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a list of entities that have liked or crushed on the specified entity.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "swipe"
                ],
                "summary": "Get incoming likes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Entity ID to get likes for",
                        "name": "entity_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Incoming likes list",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/swipe.IncomingLikeResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/swipe/likes/count": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the total count of unexpired likes and the photo of the most recent liker for the specified entity.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "swipe"
                ],
                "summary": "Get likes summary (count and last photo)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Entity ID to get likes summary for",
                        "name": "entity_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Likes summary data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/swipe.LikesSummaryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/swipe/likes/sent": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a list of entities that the specified entity has liked.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "swipe"
                ],
                "summary": "Get sent likes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Entity ID to get sent likes for",
                        "name": "entity_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sent likes list",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/swipe.SentLikeResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/swipe/unlike/{entity_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes a one-way swipe from the current active entity to another entity.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "swipe"
                ],
                "summary": "Unlike an entity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Target Entity ID",
                        "name": "entity_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Swiper Entity ID",
                        "name": "swiper_entity_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully unliked",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/swipe/unmatch/{entity_id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes an existing match between the specified swiper entity and another entity.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "swipe"
                ],
                "summary": "Unmatch an entity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Swiper Entity ID",
                        "name": "swiper_entity_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully unmatched",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
        "response.EntityResponse": {
            "type": "object",
            "properties": {
                "group": {
                    "$ref": "#/definitions/response.GroupResponse"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/response.UserResponse"
                }
            }
        },
        "response.GroupResponse": {
            "type": "object",
            "properties": {
                "created_by": {
                    "type": "string"
                },
                "entity_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "main_photos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.UserResponse"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "response.UserResponse": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "bio": {
                    "type": "string"
                },
                "consumables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.ConsumableItemResponse"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "entity_id": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/response.MasterItemResponse"
                },
                "height_cm": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "interested_genders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.MasterItemResponse"
                    }
                },
                "interests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.MasterItemResponse"
                    }
                },
                "languages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.MasterItemResponse"
                    }
                },
                "latitude": {
                    "type": "number"
                },
                "location_city": {
                    "type": "string"
                },
                "location_country": {
                    "type": "string"
                },
                "longitude": {
                    "type": "number"
                },
                "main_photo": {
                    "type": "string"
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.PhotoResponse"
                    }
                },
                "relationship_type": {
                    "$ref": "#/definitions/response.MasterItemResponse"
                },
                "status": {
                    "$ref": "#/definitions/entities.UserStatus"
                },
                "subscription": {
                    "$ref": "#/definitions/response.UserSubscriptionResponse"
                },
                "updated_at": {
                    "type": "string"
                },
                "verified_at": {
                    "type": "string"
                }
            }
        },
...
        "swipe.IncomingLikeResponse": {
            "type": "object",
            "properties": {
                "entity": {
                    "$ref": "#/definitions/response.EntityResponse"
                },
                "is_boosted": {
                    "type": "boolean"
                },
                "is_crush": {
                    "type": "boolean"
                },
                "swipe_time": {
                    "type": "string"
                },
                "target_entity_id": {
                    "type": "string"
                }
            }
        },
        "swipe.LikesSummaryResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "last_photo": {
                    "type": "string"
                }
            }
        },
        "swipe.MatchResponse": {
            "type": "object",
            "properties": {
                "is_match": {
                    "type": "boolean"
                },
                "match_id": {
                    "type": "string"
                },
                "matched_entity": {
                    "$ref": "#/definitions/response.EntityResponse"
                }
            }
        },
        "swipe.SentLikeResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "entity": {
                    "$ref": "#/definitions/response.EntityResponse"
                },
                "expires_at": {
                    "type": "string"
                },
                "is_boosted": {
                    "type": "boolean"
                },
                "is_crush": {
                    "type": "boolean"
                },
                "swiper_entity_id": {
                    "type": "string"
                }
            }
        },
        "swipe.SwipeRequest": {
            "type": "object",
            "required": [
                "direction",
                "swiped_entity_id",
                "swiper_entity_id"
            ],
            "properties": {
                "direction": {
                    "type": "string",
                    "enum": [
                        "LIKE",
                        "DISLIKE",
                        "CRUSH"
                    ]
                },
                "swiped_entity_id": {
                    "type": "string"
                },
                "swiper_entity_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type 'Bearer ' followed by your JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}